+++ code/hyperdb/src/db/main.ts
@@ -1,0 +1,1 @@
+export * from './Db';
