--- code/hyperdb/src/main.db/index.ts
@@ -1,4 +1,0 @@
-import { Db } from './Db';
-import { Swarm } from './Swarm';
-
-export { Db as HyperDb, Swarm };
