+++ code/hyperdb/src/swarm/main.ts
@@ -1,0 +1,1 @@
+export * from './Swarm';
